fix(data-table): viewport-fixed header chrome and split module
Keep the sticky header plate as wide as the table viewport instead of
content width, so horizontal scroll no longer peels header corners.
Extract pure helpers out of the oversized DataTable component file.
