+++ packages/ui/src/data-table/__tests__/DataTable.interactions.test.tsx
@@ -57,9 +57,7 @@ function DataTableHarness() {
 }
 
 function visibleRowNames() {
-  return screen
-    .getAllByTestId("row-name")
-    .map((element) => element.textContent);
+  return screen.getAllByTestId("row-name").map((element) => element.textContent);
 }
 
 describe("DataTable sorting and row reordering", () => {
@@ -74,19 +72,18 @@ describe("DataTable sorting and row reordering", () => {
       'td[data-vt-column-key="name"]',
     );
     expect(firstNameCell).toHaveClass("border-b", "data-table-test-cell");
-    expect(
-      firstNameCell?.querySelector("[data-table-cell-overflow]"),
-    ).toHaveClass("data-table-test-content");
-    expect(
-      firstNameCell?.querySelector("[data-table-cell-overflow]"),
-    ).not.toHaveClass("border-b", "data-table-test-cell");
+    expect(firstNameCell?.querySelector("[data-table-cell-overflow]")).toHaveClass(
+      "data-table-test-content",
+    );
+    expect(firstNameCell?.querySelector("[data-table-cell-overflow]")).not.toHaveClass(
+      "border-b",
+      "data-table-test-cell",
+    );
 
     await user.click(sortTrigger);
     const menu = screen.getByRole("menu");
     expect(menu).toHaveClass("min-w-28");
-    expect(screen.getByRole("menuitem", { name: "Ascending" })).toHaveClass(
-      "text-xs",
-    );
+    expect(screen.getByRole("menuitem", { name: "Ascending" })).toHaveClass("text-xs");
     await user.click(screen.getByRole("menuitem", { name: "Ascending" }));
 
     expect(visibleRowNames()).toEqual(["alpha", "bravo", "charlie"]);
@@ -137,30 +134,22 @@ describe("DataTable sorting and row reordering", () => {
     });
     fireEvent.pointerMove(window, { pointerId: 7, clientY: 115 });
 
-    const dragPreview = document.querySelector<HTMLElement>(
-      "[data-vt-row-reorder-preview]",
-    );
+    const dragPreview = document.querySelector<HTMLElement>("[data-vt-row-reorder-preview]");
     expect(dragPreview).toHaveTextContent("alpha");
     expect(dragPreview).toHaveClass("border", "border-slate-200/90");
     expect(
       Array.from(dragPreview?.querySelectorAll("td") ?? []).every(
-        (cell) =>
-          cell.style.borderTopWidth === "0px" &&
-          cell.style.borderBottomWidth === "0px",
+        (cell) => cell.style.borderTopWidth === "0px" && cell.style.borderBottomWidth === "0px",
       ),
     ).toBe(true);
-    expect(
-      document.querySelector("[data-vt-row-reorder-drop-indicator]"),
-    ).toBeNull();
+    expect(document.querySelector("[data-vt-row-reorder-drop-indicator]")).toBeNull();
     expect(tableRows[0]).toHaveStyle({ opacity: "0" });
     expect(tableRows[1]).toHaveStyle({ transform: "translate3d(0, -40px, 0)" });
     expect(tableRows[2]).toHaveStyle({ transform: "translate3d(0, -40px, 0)" });
 
     fireEvent.pointerUp(window, { pointerId: 7, clientY: 115 });
 
-    await waitFor(() =>
-      expect(visibleRowNames()).toEqual(["bravo", "charlie", "alpha"]),
-    );
+    await waitFor(() => expect(visibleRowNames()).toEqual(["bravo", "charlie", "alpha"]));
     expect(document.querySelector("[data-vt-row-reorder-preview]")).toBeNull();
     tableRows.forEach((row) => {
       expect(row.style.opacity).toBe("");
@@ -186,23 +175,14 @@ describe("DataTable column reorder handle layout", () => {
       />,
     );
 
-    const header = document.querySelector<HTMLElement>(
-      'th[data-vt-column-key="name"]',
-    );
-    const content = header?.querySelector<HTMLElement>(
-      "[data-vt-column-header-content]",
-    );
-    const handle = header?.querySelector<HTMLElement>(
-      "[data-vt-column-reorder-handle]",
-    );
+    const header = document.querySelector<HTMLElement>('th[data-vt-column-key="name"]');
+    const content = header?.querySelector<HTMLElement>("[data-vt-column-header-content]");
+    const handle = header?.querySelector<HTMLElement>("[data-vt-column-reorder-handle]");
     expect(header).not.toBeNull();
     expect(content).not.toBeNull();
     expect(handle).not.toBeNull();
     expect(content).toHaveClass("px-5");
-    expect(content).not.toHaveClass(
-      "group-hover/column:pl-5",
-      "group-focus-within/column:pl-5",
-    );
+    expect(content).not.toHaveClass("group-hover/column:pl-5", "group-focus-within/column:pl-5");
     expect(handle).toHaveClass("absolute", "left-1");
   });
 });
@@ -295,12 +275,8 @@ describe("DataTable empty state", () => {
     );
 
     const table = document.querySelector<HTMLTableElement>("table[data-vt-empty='true']");
-    const viewport = document.querySelector<HTMLElement>(
-      "[data-scrollbar-visibility='hover']",
-    );
-    const emptyRow = document.querySelector<HTMLTableRowElement>(
-      "tr[data-vt-empty-row]",
-    );
+    const viewport = document.querySelector<HTMLElement>("[data-scrollbar-visibility='hover']");
+    const emptyRow = document.querySelector<HTMLTableRowElement>("tr[data-vt-empty-row]");
 
     expect(table).not.toBeNull();
     expect(table).toHaveClass("min-w-0");
@@ -312,9 +288,7 @@ describe("DataTable empty state", () => {
     // Default EmptyState icon well (Inbox) when emptyIcon is omitted.
     expect(document.querySelector("[data-empty-icon]")).not.toBeNull();
 
-    const firstHeader = document.querySelector<HTMLElement>(
-      'th[data-vt-column-key="id"]',
-    );
+    const firstHeader = document.querySelector<HTMLElement>('th[data-vt-column-key="id"]');
     expect(firstHeader).not.toBeNull();
     expect(firstHeader?.className).not.toMatch(/min-w-\[320px\]/);
     expect(firstHeader).not.toHaveClass("sticky");
@@ -324,10 +298,7 @@ describe("DataTable empty state", () => {
 describe("DataTable scroll chrome and row dividers", () => {
   test("keeps header cells attached and forwards boundary wheel scrolling to the parent", () => {
     render(
-      <div
-        data-testid="parent-scroll"
-        style={{ height: 160, overflowY: "auto" }}
-      >
+      <div data-testid="parent-scroll" style={{ height: 160, overflowY: "auto" }}>
         <DataTable
           rows={initialRows}
           columns={plainColumns}
@@ -342,9 +313,7 @@ describe("DataTable scroll chrome and row dividers", () => {
     );
 
     const parent = screen.getByTestId("parent-scroll");
-    const viewport = document.querySelector<HTMLElement>(
-      "[data-scrollbar-visibility='hover']",
-    );
+    const viewport = document.querySelector<HTMLElement>("[data-scrollbar-visibility='hover']");
     expect(viewport).not.toBeNull();
     if (!viewport) return;
 
@@ -365,12 +334,15 @@ describe("DataTable scroll chrome and row dividers", () => {
     expect(parent.scrollTop).toBe(120);
     expect(viewport).toHaveClass("overscroll-y-auto");
     expect(viewport).not.toHaveClass("overscroll-y-none");
-    expect(document.querySelector("[data-vt-header-chrome]")).not.toBeNull();
+    const headerChrome = document.querySelector("[data-vt-header-chrome]");
+    expect(headerChrome).not.toBeNull();
+    expect(headerChrome).toHaveClass("absolute", "left-0", "top-0");
 
     const headerCells = Array.from(document.querySelectorAll("thead th"));
     expect(headerCells).not.toHaveLength(0);
     headerCells.forEach((cell) => {
-      expect(cell).toHaveClass("sticky", "top-0", "bg-slate-100");
+      // Free-scroll headers stay transparent; viewport chrome owns the plate fill.
+      expect(cell).toHaveClass("sticky", "top-0", "bg-transparent");
     });
   });
 
@@ -405,11 +377,7 @@ describe("DataTable scroll chrome and row dividers", () => {
       });
     });
     Array.from(rows.at(-1)?.cells ?? []).forEach((cell) => {
-      expect(cell).not.toHaveClass(
-        "border-b",
-        "first:rounded-l-lg",
-        "last:rounded-r-lg",
-      );
+      expect(cell).not.toHaveClass("border-b", "first:rounded-l-lg", "last:rounded-r-lg");
     });
   });
 
@@ -455,13 +423,16 @@ describe("DataTable scroll chrome and row dividers", () => {
 
     const selectHeader = document.querySelector('th[data-vt-column-key="select"]');
     const nameHeader = document.querySelector('th[data-vt-column-key="name"]');
-    const actionsHeader = document.querySelector(
-      'th[data-vt-column-key="actions"]',
-    );
+    const actionsHeader = document.querySelector('th[data-vt-column-key="actions"]');
     const headerGutter = document.querySelector("[data-vt-header-gutter]");
-    expect(selectHeader).toHaveClass("rounded-l-xl", "bg-slate-100");
+    // Sticky cells stay opaque; free middle headers stay transparent over chrome.
+    // Corner radius lives on the viewport chrome (and gutter), not per th.
+    expect(selectHeader).toHaveClass("bg-slate-100");
+    expect(selectHeader).not.toHaveClass("rounded-l-xl");
     expect(nameHeader).toHaveClass("bg-transparent");
-    expect(actionsHeader).toHaveClass("rounded-r-xl");
+    expect(actionsHeader).toHaveClass("bg-slate-100");
+    expect(actionsHeader).not.toHaveClass("rounded-r-xl");
+    expect(headerChrome).toHaveClass("absolute", "left-0", "top-0");
     // With a vertical gutter, chrome is left-only; otherwise full rounded-xl.
     if (headerGutter) {
       expect(headerGutter).toHaveClass("rounded-r-xl");
