+++ packages/ui/src/data-table/tableStorage.ts
@@ -1,139 +1,43 @@
 import type { DataTableColumn } from "./DataTable.types";
+import {
+  COLUMN_ORDER_STORAGE_PREFIX,
+  COLUMN_RESIZE_DEBUG_STORAGE_KEY,
+  COLUMN_WIDTH_STORAGE_PREFIX,
+  DEFAULT_MAX_COLUMN_WIDTH,
+  DEFAULT_MIN_COLUMN_WIDTH,
+  NON_RESIZABLE_COLUMN_KEYS,
+  TAILWIND_SPACING_UNIT_PX,
+  type ColumnOrder,
+  type ColumnWidthMap,
+} from "./dataTableModel";
 
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-export const DEFAULT_ROW_HEIGHT = 44;
-export const DEFAULT_OVERSCAN = 12;
-export const DEFAULT_SCROLL_THRESHOLD = 100;
-export const DEFAULT_BOTTOM_DEBOUNCE_MS = 120;
-export const COLUMN_WIDTH_STORAGE_PREFIX = "codeProxy.dataTable.columnWidths.v1";
-export const COLUMN_RESIZE_DEBUG_STORAGE_KEY = "codeProxy.dataTable.debugResize";
-export const DEFAULT_MIN_COLUMN_WIDTH = 72;
-export const DEFAULT_MAX_COLUMN_WIDTH = 640;
-export const COLUMN_RESIZE_PREVIEW_LINE_WIDTH = 2;
-export const NON_RESIZABLE_COLUMN_KEYS = new Set(["select", "action", "actions"]);
-const TAILWIND_SPACING_UNIT_PX = 4;
+export {
+  DEFAULT_ROW_HEIGHT,
+  DEFAULT_OVERSCAN,
+  DEFAULT_SCROLL_THRESHOLD,
+  DEFAULT_BOTTOM_DEBOUNCE_MS,
+  COLUMN_WIDTH_STORAGE_PREFIX,
+  COLUMN_RESIZE_DEBUG_STORAGE_KEY,
+  DEFAULT_MIN_COLUMN_WIDTH,
+  DEFAULT_MAX_COLUMN_WIDTH,
+  COLUMN_RESIZE_PREVIEW_LINE_WIDTH,
+  NON_RESIZABLE_COLUMN_KEYS,
+  COLUMN_ORDER_STORAGE_PREFIX,
+  COLUMN_REORDER_ACTIVATION_DELAY_MS,
+  COLUMN_REORDER_MIN_DRAG_DISTANCE_PX,
+  NON_REORDERABLE_COLUMN_KEYS,
+  type ColumnOrder,
+  type ColumnWidthMap,
+  type ColumnResizeState,
+  type ColumnResizePreview,
+  type ScrollMetrics,
+} from "./dataTableModel";
 
-// ---------------------------------------------------------------------------
-// Column Reorder
-// ---------------------------------------------------------------------------
-export const COLUMN_ORDER_STORAGE_PREFIX = "codeProxy.dataTable.columnOrder.v1";
-export const COLUMN_REORDER_ACTIVATION_DELAY_MS = 180;
-export const COLUMN_REORDER_MIN_DRAG_DISTANCE_PX = 4;
-export const NON_REORDERABLE_COLUMN_KEYS = new Set(["select", "action", "actions"]);
-
-export type ColumnOrder = string[];
-
-export interface ColumnReorderState {
-  pointerId: number;
-  columnKey: string;
-  originIndex: number;
-  currentIndex: number;
-  startClientX: number;
-  startClientY: number;
-  activated: boolean;
-  activationTimer: number | null;
-  allowedMinIndex: number;
-  allowedMaxIndex: number;
-}
-
-export interface ColumnReorderPreview {
-  fromIndex: number;
-  toIndex: number;
-  left: number;
-  top: number;
-  height: number;
-}
-
-export type ColumnWidthMap = Record<string, number>;
-
-export interface ColumnResizeState {
-  pointerId: number;
-  columnKey: string;
-  startLeftClientX: number;
-  pointerBoundaryOffsetClientX: number;
-  minWidth: number;
-  maxWidth: number;
-  previewTop: number;
-  previewBottom: number;
-  currentWidth: number;
-  lastDebugAtMs: number;
-  debugEnabled: boolean;
-}
-
-export interface ColumnResizePreview {
-  width: number;
-  left: number;
-  top: number;
-  height: number;
-  tooltipTop: number;
-}
-
-export interface ScrollMetrics {
-  scrollTop: number;
-  scrollLeft: number;
-  scrollHeight: number;
-  scrollWidth: number;
-  clientHeight: number;
-  clientWidth: number;
-}
-
-export function hasHorizontalOverflow(metrics: ScrollMetrics) {
-  return metrics.scrollWidth > metrics.clientWidth + 1;
-}
-
-export function hasVerticalOverflow(metrics: ScrollMetrics, headerHeight: number) {
-  const effectiveViewportY = Math.max(0, metrics.clientHeight - headerHeight);
-  const effectiveContentY = Math.max(effectiveViewportY, metrics.scrollHeight - headerHeight);
-  return effectiveContentY > effectiveViewportY + 1;
-}
-
-export function calculateScrollbarThumbs(scrollMetrics: ScrollMetrics, headerHeight: number) {
-  const trackInset = 8; // matches `inset-y-2` / `inset-x-2` (8px)
-  const effectiveViewportY = Math.max(0, scrollMetrics.clientHeight - headerHeight);
-  const effectiveContentY = Math.max(effectiveViewportY, scrollMetrics.scrollHeight - headerHeight);
-  const hasV = hasVerticalOverflow(scrollMetrics, headerHeight);
-  const hasH = hasHorizontalOverflow(scrollMetrics);
-
-  const v = (() => {
-    if (!hasV) return null;
-    const trackLength = Math.max(0, scrollMetrics.clientHeight - headerHeight - trackInset * 2);
-    const viewport = Math.max(1, effectiveViewportY);
-    const content = Math.max(viewport, effectiveContentY);
-    const thumbLength = Math.max(28, Math.round((viewport / content) * trackLength));
-    const maxThumbOffset = Math.max(0, trackLength - thumbLength);
-    const scrollRange = Math.max(1, scrollMetrics.scrollHeight - scrollMetrics.clientHeight);
-    const offset = Math.min(
-      maxThumbOffset,
-      Math.max(0, Math.round((scrollMetrics.scrollTop / scrollRange) * maxThumbOffset)),
-    );
-    return { top: offset, height: thumbLength };
-  })();
-
-  const h = (() => {
-    if (!hasH) return null;
-    const trackLength = Math.max(0, scrollMetrics.clientWidth - trackInset * 2);
-    const viewport = scrollMetrics.clientWidth;
-    const content = scrollMetrics.scrollWidth;
-    const thumbLength = Math.max(28, Math.round((viewport / content) * trackLength));
-    const maxThumbOffset = Math.max(0, trackLength - thumbLength);
-    const scrollRange = Math.max(1, content - viewport);
-    const offset = Math.min(
-      maxThumbOffset,
-      Math.max(0, Math.round((scrollMetrics.scrollLeft / scrollRange) * maxThumbOffset)),
-    );
-    return { left: offset, width: thumbLength };
-  })();
-
-  return { vThumb: v, hThumb: h };
-}
-
-export function clampColumnWidth<T>(column: DataTableColumn<T>, width: number) {
-  const minWidth = resolveColumnMinWidth(column);
-  const maxWidth = resolveColumnMaxWidth(column, minWidth);
-  return Math.max(minWidth, Math.min(maxWidth, Math.round(width)));
-}
+export {
+  hasHorizontalOverflow,
+  hasVerticalOverflow,
+  calculateScrollbarThumbs,
+} from "./scrollMetrics";
 
 function parseTailwindSizePx(token: string) {
   const arbitrary = token.match(/^\[(\d+(?:\.\d+)?)(px|rem)\]$/);
@@ -142,9 +46,7 @@ function parseTailwindSizePx(token: string) {
     if (!Number.isFinite(value)) return null;
     return arbitrary[2] === "rem" ? Math.round(value * 16) : Math.round(value);
   }
-
   if (token === "px") return 1;
-
   const numeric = Number(token);
   if (!Number.isFinite(numeric)) return null;
   return Math.round(numeric * TAILWIND_SPACING_UNIT_PX);
@@ -162,7 +64,9 @@ function resolveWidthClassPx(width: string | undefined, prefix: string) {
 }
 
 export function resolveColumnMinWidth<T>(column: DataTableColumn<T>) {
-  return column.minWidthPx ?? resolveWidthClassPx(column.width, "min-w") ?? DEFAULT_MIN_COLUMN_WIDTH;
+  return (
+    column.minWidthPx ?? resolveWidthClassPx(column.width, "min-w") ?? DEFAULT_MIN_COLUMN_WIDTH
+  );
 }
 
 export function resolveColumnMaxWidth<T>(
@@ -170,6 +74,12 @@ export function resolveColumnMaxWidth<T>(
   minWidth = resolveColumnMinWidth(column),
 ) {
   return Math.max(minWidth, column.maxWidthPx ?? DEFAULT_MAX_COLUMN_WIDTH);
+}
+
+export function clampColumnWidth<T>(column: DataTableColumn<T>, width: number) {
+  const minWidth = resolveColumnMinWidth(column);
+  const maxWidth = resolveColumnMaxWidth(column, minWidth);
+  return Math.max(minWidth, Math.min(maxWidth, Math.round(width)));
 }
 
 export function getColumnWidthStorageKey(tableId?: string) {
@@ -180,13 +90,11 @@ export function getColumnWidthStorageKey(tableId?: string) {
 export function readStoredColumnWidths(tableId?: string): ColumnWidthMap {
   const key = getColumnWidthStorageKey(tableId);
   if (!key || typeof window === "undefined") return {};
-
   try {
     const raw = window.localStorage.getItem(key);
     if (!raw) return {};
     const parsed = JSON.parse(raw) as unknown;
     if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) return {};
-
     return Object.fromEntries(
       Object.entries(parsed as Record<string, unknown>)
         .filter(([, value]) => typeof value === "number" && Number.isFinite(value))
@@ -200,7 +108,6 @@ export function readStoredColumnWidths(tableId?: string): ColumnWidthMap {
 export function writeStoredColumnWidths(tableId: string | undefined, widths: ColumnWidthMap) {
   const key = getColumnWidthStorageKey(tableId);
   if (!key || typeof window === "undefined") return;
-
   try {
     const normalized = Object.fromEntries(
       Object.entries(widths).filter(([, value]) => Number.isFinite(value) && value > 0),
@@ -234,12 +141,10 @@ export function shouldAllowColumnResize<T>(
 
 export function resolveCellOverflowTooltip<T>(column: DataTableColumn<T>, row: T, index: number) {
   if (column.overflowTooltip === false) return false;
-
   if (typeof column.overflowTooltip === "function") {
     const value = column.overflowTooltip(row, index);
     return value === null || value === undefined ? null : String(value);
   }
-
   return undefined;
 }
 
@@ -253,12 +158,12 @@ export function safeSetPointerCapture(element: Element, pointerId: number) {
   }
 }
 
-function getColumnOrderStorageKey(tableId?: string) {
+export function getColumnOrderStorageKey(tableId?: string) {
   const trimmed = tableId?.trim();
   return trimmed ? `${COLUMN_ORDER_STORAGE_PREFIX}.${trimmed}` : null;
 }
 
-function readStoredColumnOrder(tableId?: string): ColumnOrder {
+export function readStoredColumnOrder(tableId?: string): ColumnOrder {
   const key = getColumnOrderStorageKey(tableId);
   if (!key || typeof window === "undefined") return [];
   try {
@@ -274,25 +179,20 @@ function readStoredColumnOrder(tableId?: string): ColumnOrder {
   }
 }
 
-function writeStoredColumnOrder(tableId: string | undefined, order: ColumnOrder) {
+export function writeStoredColumnOrder(tableId: string | undefined, order: ColumnOrder) {
   const key = getColumnOrderStorageKey(tableId);
   if (!key || typeof window === "undefined") return;
   try {
     const normalized = Array.from(new Set(order.filter((value) => value.trim() !== "")));
     window.localStorage.setItem(key, JSON.stringify(normalized));
-  } catch {}
+  } catch {
+    // localStorage can be unavailable in private browsing or embedded contexts.
+  }
 }
 
-function resolveColumnOrderLock<T>(column: DataTableColumn<T>) {
+export function resolveColumnOrderLock<T>(column: DataTableColumn<T>) {
   if (column.lockOrder) return column.lockOrder;
   if (column.key === "select") return "start";
   if (column.key === "action" || column.key === "actions") return "end";
   return null;
 }
-
-export {
-  getColumnOrderStorageKey,
-  readStoredColumnOrder,
-  writeStoredColumnOrder,
-  resolveColumnOrderLock,
-};
